+++ src/ibex_bluesky_core/callbacks/reflectometry/_det_map.py
@@ -33,7 +33,7 @@ from event_model import Event, EventDescriptor, RunStart, RunStop
 from matplotlib.axes import Axes
 
 from ibex_bluesky_core.callbacks import show_plot
-from ibex_bluesky_core.devices.simpledae import VARIANCE_ADDITION
+from ibex_bluesky_core.devices.dae.strategies import VARIANCE_ADDITION
 
 logger = logging.getLogger(__name__)
