+++ src/ibex_bluesky_core/devices/dae/strategies/_base.py
@@ -1,13 +1,7 @@
 """Base classes for DAE strategies."""
+from ophyd_async.core import Device
 
-from typing import TYPE_CHECKING
-
-from ophyd_async.core import (
-    Device,
-)
-
-if TYPE_CHECKING:
-    from ibex_bluesky_core.devices.simpledae import SimpleDae
+from ibex_bluesky_core.devices.dae import Dae
 
 
 class ProvidesExtraReadables:
@@ -16,7 +10,7 @@ class ProvidesExtraReadables:
     Those signals will then be added to read() and describe() on the top-level SimpleDae object.
     """
 
-    def additional_readable_signals(self, dae: "SimpleDae") -> list[Device]:
+    def additional_readable_signals(self, dae: Dae) -> list[Device]:
         """Define signals that this strategy considers important.
 
         These will be added to the dae's default-read signals and made available by read() on the
@@ -28,30 +22,30 @@ class ProvidesExtraReadables:
 class Controller(ProvidesExtraReadables):
     """Controller specifies how DAE runs should be started & stopped."""
 
-    async def start_counting(self, dae: "SimpleDae") -> None:
+    async def start_counting(self, dae: Dae) -> None:
         """Start counting for a single scan point."""
 
-    async def stop_counting(self, dae: "SimpleDae") -> None:
+    async def stop_counting(self, dae: Dae) -> None:
         """Stop counting for a single scan point."""
 
-    async def setup(self, dae: "SimpleDae") -> None:
+    async def setup(self, dae: Dae) -> None:
         """Pre-scan setup."""
 
-    async def teardown(self, dae: "SimpleDae") -> None:
+    async def teardown(self, dae: Dae) -> None:
         """Post-scan teardown."""
 
 
 class Waiter(ProvidesExtraReadables):
     """Waiter specifies how the dae will wait for a scan point to complete counting."""
 
-    async def wait(self, dae: "SimpleDae") -> None:
+    async def wait(self, dae: Dae) -> None:
         """Wait for the acquisition to complete."""
 
 
 class Reducer(ProvidesExtraReadables):
     """Reducer specifies any post-processing which needs to be done after a scan point completes."""
 
-    async def reduce_data(self, dae: "SimpleDae") -> None:
+    async def reduce_data(self, dae: Dae) -> None:
         """Triggers a reduction of DAE data after a scan point has been measured.
 
         Data that should be published by this reducer should be added as soft signals, in
