+++ src/ibex_bluesky_core/devices/dae/strategies/_waiters.py
@@ -3,7 +3,7 @@
 import asyncio
 import logging
 from abc import ABC, abstractmethod
-from typing import TYPE_CHECKING, Generic, TypeVar
+from typing import Generic, TypeVar
 
 from ophyd_async.core import (
     Device,
@@ -11,12 +11,11 @@ from ophyd_async.core import (
     wait_for_value,
 )
 
-from ibex_bluesky_core.devices.simpledae._strategies import Waiter
+from ibex_bluesky_core.devices.dae import Dae
+from ibex_bluesky_core.devices.dae.strategies._base import Waiter
+
 
 logger = logging.getLogger(__name__)
-
-if TYPE_CHECKING:
-    from ibex_bluesky_core.devices.simpledae import SimpleDae
 
 
 T = TypeVar("T", int, float)
@@ -34,26 +33,26 @@ class SimpleWaiter(Waiter, Generic[T], ABC):
         """
         self._value: T = value
 
-    async def wait(self, dae: "SimpleDae") -> None:
+    async def wait(self, dae: Dae) -> None:
         """Wait for signal to reach the user-specified value."""
         signal = self.get_signal(dae)
         logger.info("starting wait for signal %s", signal.source)
         await wait_for_value(signal, lambda v: v >= self._value, timeout=None)
         logger.info("completed wait for signal %s", signal.source)
 
-    def additional_readable_signals(self, dae: "SimpleDae") -> list[Device]:
+    def additional_readable_signals(self, dae: Dae) -> list[Device]:
         """Publish the signal we're waiting on as an interesting signal."""
         return [self.get_signal(dae)]
 
     @abstractmethod
-    def get_signal(self, dae: "SimpleDae") -> SignalR[T]:
+    def get_signal(self, dae: Dae) -> SignalR[T]:
         """Get the numeric signal to wait for."""
 
 
 class PeriodGoodFramesWaiter(SimpleWaiter[int]):
     """Wait for period good frames to reach a user-specified value."""
 
-    def get_signal(self, dae: "SimpleDae") -> SignalR[int]:
+    def get_signal(self, dae: Dae) -> SignalR[int]:
         """Wait for period good frames."""
         return dae.period.good_frames
 
@@ -61,7 +60,7 @@ class PeriodGoodFramesWaiter(SimpleWaiter[int]):
 class GoodFramesWaiter(SimpleWaiter[int]):
     """Wait for good frames to reach a user-specified value."""
 
-    def get_signal(self, dae: "SimpleDae") -> SignalR[int]:
+    def get_signal(self, dae: Dae) -> SignalR[int]:
         """Wait for good frames."""
         return dae.good_frames
 
@@ -69,7 +68,7 @@ class GoodFramesWaiter(SimpleWaiter[int]):
 class GoodUahWaiter(SimpleWaiter[float]):
     """Wait for good microamp-hours to reach a user-specified value."""
 
-    def get_signal(self, dae: "SimpleDae") -> SignalR[float]:
+    def get_signal(self, dae: Dae) -> SignalR[float]:
         """Wait for good uah."""
         return dae.good_uah
 
@@ -77,7 +76,7 @@ class GoodUahWaiter(SimpleWaiter[float]):
 class MEventsWaiter(SimpleWaiter[float]):
     """Wait for a user-specified number of millions of events."""
 
-    def get_signal(self, dae: "SimpleDae") -> SignalR[float]:
+    def get_signal(self, dae: Dae) -> SignalR[float]:
         """Wait for mevents."""
         return dae.m_events
 
@@ -94,7 +93,7 @@ class TimeWaiter(Waiter):
         """
         self._secs = seconds
 
-    async def wait(self, dae: "SimpleDae") -> None:
+    async def wait(self, dae: Dae) -> None:
         """Wait for the specified time duration."""
         logger.info("starting wait for %f seconds", self._secs)
         await asyncio.sleep(self._secs)
