+++ src/ibex_bluesky_core/devices/polarisingdae/__init__.py
@@ -1,39 +1,34 @@
 """An interface to the DAE for bluesky, suited for polarisation."""
 
 import logging
-
-import scipp as sc
-from bluesky.protocols import Movable
-from ophyd_async.core import (
-    AsyncStatus,
-    Reference,
-)
+from typing import Generic
 from typing_extensions import TypeVar
 
-from ibex_bluesky_core.devices.polarisingdae._reducers import (
+import scipp as sc
+from bluesky.protocols import Movable, Triggerable
+from ophyd_async.core import (
+    AsyncStatus,
+    Reference, AsyncStageable,
+)
+
+from ibex_bluesky_core.devices.dae import Dae
+from ibex_bluesky_core.devices.dae.strategies import (
     PolarisingReducer,
     WavelengthBoundedNormalizer,
     polarization,
-)
-from ibex_bluesky_core.devices.simpledae import (
     Controller,
-    GoodFramesWaiter,
-    PeriodGoodFramesWaiter,
-    PeriodPerPointController,
-    Reducer,
-    RunPerPointController,
-    SimpleDae,
     Waiter,
+    Reducer,
     wavelength_bounded_spectra,
+    PeriodPerPointController,
+    RunPerPointController,
+    GoodFramesWaiter,
+    PeriodGoodFramesWaiter
 )
+
 from ibex_bluesky_core.utils import get_pv_prefix
 
 logger = logging.getLogger(__name__)
-
-
-TController_co = TypeVar("TController_co", bound="Controller", default="Controller", covariant=True)
-TWaiter_co = TypeVar("TWaiter_co", bound="Waiter", default="Waiter", covariant=True)
-TReducer_co = TypeVar("TReducer_co", bound="Reducer", default="Reducer", covariant=True)
 
 __all__ = [
     "PolarisingDae",
@@ -44,7 +39,12 @@ __all__ = [
 ]
 
 
-class PolarisingDae(SimpleDae):
+TController_co = TypeVar("TController_co", bound="Controller", default=Controller, covariant=True)
+TWaiter_co = TypeVar("TWaiter_co", bound="Waiter", default=Waiter, covariant=True)
+TReducer_co = TypeVar("TReducer_co", bound="Reducer", default=Reducer, covariant=True)
+
+
+class PolarisingDae(Dae, Triggerable, AsyncStageable, Generic[TController_co, TWaiter_co, TReducer_co]):
     """DAE with strategies for data collection, waiting, and reduction, suited for polarisation.
 
     This class is a more complex version of SimpleDae, with a more complex set of strategies.
@@ -62,7 +62,7 @@ class PolarisingDae(SimpleDae):
         reducer: TReducer_co,
         reducer_up: TReducer_co,
         reducer_down: TReducer_co,
-        flipper: Movable,
+        flipper: Movable[float],
         flipper_states: tuple[float, float],
     ) -> None:
         """Initialise a DAE with polarisation strategies.
@@ -84,7 +84,7 @@ class PolarisingDae(SimpleDae):
             flipper_states: A tuple of two floats, the neutron states to be set between runs.
 
         """
-        self.flipper: Reference[Movable] = Reference(flipper)
+        self.flipper: Reference[Movable[float]] = Reference(flipper)
         self.flipper_states: tuple[float, float] = flipper_states
 
         self._prefix = prefix
@@ -108,7 +108,7 @@ class PolarisingDae(SimpleDae):
         # controller, waiter and reducers may be Devices (but don't necessarily have to be),
         # so can define their own signals. do __init__ after defining those, so that the signals
         # are connected/named and usable.
-        super(SimpleDae, self).__init__(prefix=prefix, name=name)
+        super().__init__(prefix=prefix, name=name)
 
         # Ask each defined strategy what it's interesting signals are, and ensure those signals are
         # published when the top-level SimpleDae object is read.
@@ -124,6 +124,13 @@ class PolarisingDae(SimpleDae):
         logger.info("extra readables: %s", list(extra_readables))
         self.add_readables(devices=list(extra_readables))
 
+
+    @AsyncStatus.wrap
+    async def stage(self) -> None:
+        """Pre-scan setup. Delegate to the controller."""
+        await self.controller.setup(self)
+
+
     @AsyncStatus.wrap
     async def trigger(self) -> None:
         """Take a single measurement and prepare it for later reading.
@@ -131,14 +138,14 @@ class PolarisingDae(SimpleDae):
         This waits for the acquisition and any defined reduction to be complete, such that
         after this coroutine completes, all relevant data is available via read()
         """
-        await self.flipper().set(self.flipper_states[0])
+        self.flipper().set(self.flipper_states[0])
 
         await self.controller.start_counting(self)
         await self.waiter.wait(self)
         await self.controller.stop_counting(self)
         await self.reducer_up.reduce_data(self)
 
-        await self.flipper().set(self.flipper_states[1])
+        self.flipper().set(self.flipper_states[1])
 
         await self.controller.start_counting(self)
         await self.waiter.wait(self)
@@ -148,10 +155,16 @@ class PolarisingDae(SimpleDae):
         await self.reducer.reduce_data(self)
 
 
+    @AsyncStatus.wrap
+    async def unstage(self) -> None:
+        """Post-scan teardown, delegate to the controller."""
+        await self.controller.teardown(self)
+
+
 def polarising_dae(  # noqa: PLR0913, PLR0917
     det_pixels: list[int],
     frames: int,
-    flipper: Movable,
+    flipper: Movable[float],
     flipper_states: tuple[float, float],
     intervals: list[sc.Variable],
     total_flight_path_length: sc.Variable,
