+++ tests/devices/strategies/test_reducers.py
@@ -8,17 +8,18 @@ import pytest
 import scipp as sc
 from ophyd_async.testing import get_mock_put, set_mock_value
 
-from ibex_bluesky_core.devices.simpledae import (
+from ibex_bluesky_core.devices.dae.strategies import (
     VARIANCE_ADDITION,
     GoodFramesNormalizer,
     MonitorNormalizer,
     PeriodGoodFramesNormalizer,
     PeriodSpecIntegralsReducer,
     ScalarNormalizer,
-    SimpleDae,
     tof_bounded_spectra,
     wavelength_bounded_spectra,
 )
+
+from ibex_bluesky_core.devices.simpledae import SimpleDae
 
 
 @pytest.fixture
