+++ tests/plans/test_det_map_align.py
@@ -6,12 +6,12 @@ import pytest
 from ophyd_async.core import soft_signal_rw
 from ophyd_async.testing import callback_on_mock_put, set_mock_value
 
-from ibex_bluesky_core.devices.simpledae import (
+from ibex_bluesky_core.devices.dae.strategies import (
     PeriodPerPointController,
     PeriodSpecIntegralsReducer,
-    SimpleDae,
     Waiter,
 )
+from ibex_bluesky_core.devices.simpledae import SimpleDae
 from ibex_bluesky_core.plans.reflectometry import (
     angle_scan_plan,
     height_and_angle_scan_plan,
